#194: address review — drop container flow (banner via tracker+ownDeviceType), extract IosViewControllerHolder, fix log tags
