+++ composeApp/src/appleMain/kotlin/com/tubetoast/tether/di/AppleAppContainer.kt
@@ -10,10 +10,8 @@ import com.tubetoast.tether.discovery.MdnsDiscovery
 import com.tubetoast.tether.identity.DataStoreFingerprintPersistence
 import com.tubetoast.tether.identity.FingerprintPersistence
 import com.tubetoast.tether.network.AppleUploadStorageBackend
-import com.tubetoast.tether.network.DefaultTransferActivityTracker
 import com.tubetoast.tether.network.FileServer
 import com.tubetoast.tether.network.FileUploadStorage
-import com.tubetoast.tether.network.TransferActivityTracker
 import com.tubetoast.tether.network.UploadStorage
 import com.tubetoast.tether.preferences.DefaultFileTransferPreferences
 import com.tubetoast.tether.preferences.FileTransferPreferences
@@ -26,7 +24,6 @@ import kotlinx.cinterop.alloc
 import kotlinx.cinterop.memScoped
 import kotlinx.cinterop.ptr
 import kotlinx.cinterop.value
-import kotlinx.coroutines.flow.StateFlow
 import okio.Path.Companion.toPath
 import platform.Foundation.NSApplicationSupportDirectory
 import platform.Foundation.NSDocumentDirectory
@@ -34,10 +31,6 @@ import platform.Foundation.NSError
 import platform.Foundation.NSFileManager
 import platform.Foundation.NSSearchPathForDirectoriesInDomains
 import platform.Foundation.NSUserDomainMask
-import platform.UIKit.UIApplication
-import platform.UIKit.UIViewController
-import platform.UIKit.UIWindow
-import platform.UIKit.UIWindowScene
 
 open class AppleAppContainer(
     private val config: AppleAppConfig,
@@ -76,40 +69,11 @@ open class AppleAppContainer(
     )
     override val ownDeviceType: DeviceType = DeviceType.Ios
 
-    override val transferActivityTracker: TransferActivityTracker by lazy { DefaultTransferActivityTracker(appScope) }
-    override val transferActiveForBanner: StateFlow<Boolean> = transferActivityTracker.active
-
-    // The hosting ComposeUIViewController, attached by MainViewController once created. Presenting
-    // from it (via its window's root) is reliable; the connectedScenes lookup is a last resort
-    // because SwiftUI-managed scenes do not always expose a key window at pick time.
-    private var hostViewController: UIViewController? = null
-
-    fun attachHostViewController(viewController: UIViewController) {
-        hostViewController = viewController
-    }
+    val iosViewControllerHolder = IosViewControllerHolder()
 
     override val filePicker: FilePicker = IosFilePicker(
-        viewControllerProvider = { topmostPresenter() },
+        viewControllerProvider = iosViewControllerHolder::topmostPresenter,
     )
-
-    private fun topmostPresenter(): UIViewController? {
-        val root = hostViewController?.view?.window?.rootViewController
-            ?: hostViewController
-            ?: connectedScenesRoot()
-        var top = root
-        while (top?.presentedViewController != null) {
-            top = top.presentedViewController
-        }
-        return top
-    }
-
-    private fun connectedScenesRoot(): UIViewController? =
-        UIApplication.sharedApplication.connectedScenes
-            .filterIsInstance<UIWindowScene>()
-            .firstNotNullOfOrNull { scene ->
-                scene.keyWindow
-                    ?: scene.windows.filterIsInstance<UIWindow>().firstOrNull { it.isKeyWindow() }
-            }?.rootViewController
 }
 
 @OptIn(ExperimentalForeignApi::class, kotlinx.cinterop.BetaInteropApi::class)
