+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/presentation/banners/BannersComponent.kt
@@ -2,7 +2,9 @@ package com.tubetoast.tether.presentation.banners
 
 import com.arkivanov.decompose.ComponentContext
 import com.arkivanov.essenty.lifecycle.coroutines.coroutineScope
+import com.tubetoast.tether.network.TransferActivityTracker
 import com.tubetoast.tether.peer.PeersRepository
+import com.tubetoast.tether.protocol.DeviceType
 import com.tubetoast.tether.transfer.PeerIdentity
 import com.tubetoast.tether.transfer.PeerTransferEngineRegistry
 import com.tubetoast.tether.transfer.PeerTransferState
@@ -18,6 +20,7 @@ import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
@@ -29,7 +32,8 @@ class BannersComponent(
     private val peersRepository: PeersRepository,
     private val engineRegistry: PeerTransferEngineRegistry,
     private val conflictRelay: PeerConflictRelay,
-    transferActive: StateFlow<Boolean>,
+    transferActivityTracker: TransferActivityTracker,
+    ownDeviceType: DeviceType,
     coroutineScope: CoroutineScope = componentContext.coroutineScope(),
 ) : ComponentContext by componentContext {
     private val scope = coroutineScope
@@ -37,7 +41,9 @@ class BannersComponent(
     val dropFeedback: StateFlow<Boolean> get() = _dropFeedback
     private val _dropFeedback = MutableStateFlow(false)
 
-    val showForegroundConstraint: StateFlow<Boolean> = transferActive
+    val showForegroundConstraint: StateFlow<Boolean> = transferActivityTracker.active
+        .map { it && ownDeviceType == DeviceType.Ios }
+        .stateIn(scope, SharingStarted.Eagerly, false)
 
     private val selectedConflictPeer = MutableStateFlow<PeerIdentity?>(null)
