+++ composeApp/src/commonTest/kotlin/com/tubetoast/tether/presentation/RootComponentTest.kt
@@ -8,6 +8,7 @@ import com.arkivanov.essenty.statekeeper.StateKeeperDispatcher
 import com.tubetoast.tether.config.DeviceNameStore
 import com.tubetoast.tether.config.EphemeralDeviceNamePersistence
 import com.tubetoast.tether.discovery.FakeDeviceDiscovery
+import com.tubetoast.tether.network.NoOpTransferActivityTracker
 import com.tubetoast.tether.peer.PeersRepository
 import com.tubetoast.tether.preferences.FakeFileTransferPreferences
 import com.tubetoast.tether.preferences.FakePeerPreferencesStore
@@ -16,6 +17,7 @@ import com.tubetoast.tether.presentation.banners.PeerConflictRelay
 import com.tubetoast.tether.presentation.devicename.DeviceNameComponent
 import com.tubetoast.tether.presentation.transfer.PeerTransferComponent
 import com.tubetoast.tether.protocol.Device
+import com.tubetoast.tether.protocol.DeviceType
 import com.tubetoast.tether.transfer.FakeFilePicker
 import com.tubetoast.tether.transfer.FakeFileSource
 import com.tubetoast.tether.transfer.PeerTransferEngine
@@ -354,7 +356,8 @@ class RootComponentTest {
                             peersRepository = peersRepository,
                             engineRegistry = fakePeerTransferEngineRegistry(coroutineScope),
                             conflictRelay = PeerConflictRelay(),
-                            transferActive = MutableStateFlow(false),
+                            transferActivityTracker = NoOpTransferActivityTracker,
+                            ownDeviceType = DeviceType.Android,
                             coroutineScope = coroutineScope,
                         )
                     },
